requirements.txt
pandas
scikit-learn
numpy
matplotlib
kneed
joblib
PyYAML